Fix issues building with Lua and MacPorts
Some users who happen to have Lua installed system-wide and/or have
MacPorts enabled may see compile failures. This commit includes a
small patch to the Redis Makefile that adjusts where CFLAGS gets
inserted into the compile command to prevent these problems.

This is a redo of #16, which got reverted because `asdf` apparently
does not set `$ASDF_PLUGIN_PATH` itself, though it does respect it
when searching for plugins.